requirements.txt
fastapi
uvicorn
redis
pydantic
python-dotenv
groq
